requirements.txt
fastapi>=0.115.6,<1.0.0
httpx>=0.28.1,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
langchain==1.3.0
langchain-core==1.4.0
langchain-google-genai==4.2.2
langchain-openai==1.1.3
langgraph==1.2.0
langgraph-checkpoint-sqlite==3.1.0
google-api-python-client==2.187.0
google-auth==2.52.0
python-dotenv==1.2.1
cachetools==5.4.0
streamlit==1.44.1
requests==2.32.3
pydantic>=2.10.0,<3.0.0
